requirements.txt
pymongo
jieba
stanfordcorenlp
opencc-python-reimplemented
